moved keycloak.ts to middleware/auth folder
