+++ src/pages/profile.tsx
@@ -1,4 +1,4 @@
-import { securedProps } from 'common/util/keycloak'
+import { securedProps } from 'middleware/auth/keycloak'
 import ProfilePage from 'components/auth/profile/ProfilePage'
 
 export const getServerSideProps = securedProps
